Thread fix
When shutting down the server, it waits on all threads to finish then it shut downs. Also fixed
wording on the shutdown html to clarify that the server has shut down and interpretation is
left for none other than to not wait for a signal that it has indeed shut down.